+++ projects/swimlane/ngx-ui/src/lib/components/select/contains-filter.util.ts
@@ -9,12 +9,10 @@ export function containsFilter(
   if (value === undefined || value === null || depth > 2) {
     return false;
   }
-
+  if (typeof value === 'number') {
+    value = String(value);
+  }
   if (typeof value === 'string') {
-    if (!isNaN(+value)) {
-      return value === keyword;
-    }
-
     const escapedKeyword = escapeRegExp(keyword);
     // eslint-disable-next-line
     const idx = options.filterCaseSensitive ? value.indexOf(keyword) : value.search(new RegExp(escapedKeyword, 'i'));