+++ projects/swimlane/ngx-ui/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fix (`ngx-date-range-picker`): Preset values now retain their relative expressions when cancel is clicked, instead of being converted to timestamps
 - Fix (`ngx-date-range-picker`): Fix highlight the date for given `selectedRange` values.
+- Fix (`ngx-select`): Fix filtering of numeric option values.
 
 ## 50.1.3 (2025-09-23)